rename NewConnectionScreen, remove old screens
