+++ BrightID/src/components/PendingConnectionsScreens/RNCameraProvider.e2e.js
