+++ BrightID/src/components/PendingConnectionsScreens/RNCameraProvider.js
