+++ BrightID/src/components/PendingConnectionsScreens/actions/notificationToken.js
