+++ BrightID/src/components/PendingConnectionsScreens/channelSlice.js
