+++ BrightID/src/reducer/index.js
@@ -1,11 +1,11 @@
 import { combineReducers } from 'redux';
 import apps from './apps';
-import channels from '../components/NewConnectionsScreens/channelSlice';
+import channels from '../components/PendingConnectionsScreens/channelSlice';
 import connections from './connections';
 import groups from './groups';
 import notifications from './notifications';
 import operations from './operations';
-import pendingConnections from '../components/NewConnectionsScreens/pendingConnectionSlice';
+import pendingConnections from '../components/PendingConnectionsScreens/pendingConnectionSlice';
 import recoveryData from './recoveryData';
 import user from './user';
