+++ BrightID/src/routes/NewConnections.js
@@ -3,11 +3,10 @@ import {
   createStackNavigator,
   CardStyleInterpolators,
 } from '@react-navigation/stack';
-import MyCodeScreen from '@/components/NewConnectionsScreens/MyCodeScreen';
-import ScanCodeScreen from '@/components/NewConnectionsScreens/ScanCodeScreen';
-import SuccessScreen from '@/components/NewConnectionsScreens/SuccessScreen';
-import PendingConnectionsScreen from '@/components/NewConnectionsScreens/PendingConnectionsScreen';
-import GroupConnectionScreen from '@/components/NewConnectionsScreens/GroupConnectionScreen';
+import MyCodeScreen from '@/components/PendingConnectionsScreens/MyCodeScreen';
+import ScanCodeScreen from '@/components/PendingConnectionsScreens/ScanCodeScreen';
+import PendingConnectionsScreen from '@/components/PendingConnectionsScreens/PendingConnectionsScreen';
+import GroupConnectionScreen from '@/components/PendingConnectionsScreens/GroupConnectionScreen';
 import { NavHome, headerOptions } from './helpers';
 
 const Stack = createStackNavigator();
@@ -40,11 +39,6 @@ const NewConnections = () => (
       name="ScanCode"
       component={ScanCodeScreen}
       options={newConnectionOptions}
-    />
-    <Stack.Screen
-      name="ConnectSuccess"
-      component={SuccessScreen}
-      options={connectionPreviewOptions}
     />
     <Stack.Screen
       name="PendingConnections"
